Delete NIDAP-specific hawk.lock and meta.yaml files; migrate to no dill usage anywhere and saving of only data objects plus breaking down and building up complex objects such as custom objects and lazyframes. Should prevent recent issues with loading archives with dill causing old code to get loaded, plus should enable executing jobs without having to save all intermediate data in the inputs, i.e., can send in lazyframes and not an entire copy of the relevant inputs. Example in running spatial UMAP shows how we must slightly modify the inputs dictionary to be compatible with the expected dictionary format for serialization/deserialization. Likewise add a wrapper function in main.py to call spatial UMAP code using this proper input/output format. Also try commenting out forced lf.collect_schema() to register index columns, though doubtful that will work and we really will have to include that to address the lazyframe column registration bug.

+++ source/framework/analysis_framework.py
@@ -86,7 +86,8 @@ def delete_serialized_files(dict_name, directory):
         os.remove(pkl_file)
 
         dill_file = os.path.join(directory, f'{dict_name}.dill')
-        os.remove(dill_file)
+        if os.path.exists(dill_file):
+            os.remove(dill_file)
 
         return True
     except Exception as e:
